[fix] TeamJoin & Leave 시 채팅 선택/ 토픽 초대 화면에서 액션 수정

+++ app/src/main/java/com/tosslab/jandi/app/services/socket/JandiSocketServiceModel.java
@@ -594,7 +594,7 @@ public class JandiSocketServiceModel {
                     getObject(object, SocketTopicPushEvent.class);
 
             SocketTopicPushEvent.Data data = socketTopicPushEvent.getData();
-            int roomId = data.getRoomId();
+            long roomId = data.getRoomId();
             boolean subscribe = data.isSubscribe();
             JandiPreference.setSocketConnectedLastTime(socketTopicPushEvent.getTs());
             TopicRepository.getInstance().updatePushSubscribe(roomId, subscribe);
